Fixed vital syntax error
Before, the extension would always interpret the check for if the website was IMDb TV as true, because I forgot to add parentheses after the function. That's fixed now. Made a couple of changes to the popup page for increased accessibility. Also starting to test support for Youtube and Vimeo.