Remove unneeded project folders
I initially thought that I needed to include the source code from the 2 other repositories I derived the code from, but after looking closer at the GPL, it appears that linking to the 2 repositories both on my repository and in the official release (once we get to that point) will be sufficient, as long as the repositories remain available on Github. Now that the ReadMe links to the 2 repositories, the folders have been removed from this repository, reducing the repository's size significantly.