Fix durationDifference issue
Before, the check for the duration difference only worked once every time because the interval was being cleared, but that is now fixed (though there may be a way to make it query only as much as needed). Also separate the check for advertisements into a different function. Also add a check for IMDb TV when the time indicator element doesn't exist yet