Remove obsolete comment
When the comment was added a couple of days ago, I had asked permission from a developer who created a Hulu ad detector extension to use some lines of code (that I was planning on shortening with short circuit evaluation). However, it appears that the Hulu video page structure has changed since the extension's last update and I have not heard from the developer, so I will not be using that code. I forgot to remove the comment until now